# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: C7NTAX API
  description: |
    Professional Services Automation API — programmatic access to the C7NTAX platform.
    All endpoints require Bearer token authentication (JWT) obtained via `/api/auth/login`.
    External apps and user-created plugins use this same API.

    ## Authentication
    1. POST `/api/auth/login` with email + password
    2. If MFA is enabled, receive `mfaToken` and POST `/api/auth/verify-mfa` with TOTP code
    3. Use returned `token` as `Authorization: Bearer <token>` header

    ## Rate Limiting
    100 requests per 15-minute window per token. 429 responses include `Retry-After`.

    ## Permissions
    Each endpoint lists required permission(s). Client users are scoped to their company's data.
  version: "1.0.0"
  contact:
    name: Cyber 7 Group
    email: support@cyber7group.com
servers:
  - url: http://localhost:4000/api
    description: Local development
  - url: https://api.C7NTAX.com/api
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /boards:
    get:
      tags: [Boards]
      summary: List all service boards
      security: [{ bearerAuth: [] }]
      responses:
        "200": { content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Board" } } } } }
    post:
      tags: [Boards]
      summary: Create a new service board
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        "201": { content: { application/json: { schema: { $ref: "#/components/schemas/Board" } } } }
  /boards/{id}:
    get:
      tags: [Boards]
      summary: Get board with connector details
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/BoardDetail" } } } }
    patch:
      tags: [Boards]
      summary: Update board settings
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { content: { application/json: { schema: { $ref: "#/components/schemas/Board" } } } }
  /boards/{boardId}/connectors:
    get:
      tags: [Boards]
      summary: List email connectors for a board
      security: [{ bearerAuth: [] }]
      parameters: [{ name: boardId, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/EmailConnector" } } } } }
    post:
      tags: [Boards]
      summary: Create email connector for ticket ingestion
      security: [{ bearerAuth: [] }]
      parameters: [{ name: boardId, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, host, user, password]
              properties:
                email: { type: string, format: email }
                host: { type: string }
                port: { type: integer, default: 993 }
                secure: { type: boolean, default: true }
                folder: { type: string, default: "INBOX" }
                user: { type: string }
                password: { type: string }
                pollIntervalSeconds: { type: integer, default: 60 }
      responses:
        "201": { content: { application/json: { schema: { $ref: "#/components/schemas/EmailConnector" } } } }

  # ─── Clients ───
  /clients:
    get:
      tags: [Clients]
      summary: List all client companies
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Paginated client list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Company" } }
                  total: { type: integer }
    post:
      tags: [Clients]
      summary: Create a new client company
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                email: { type: string }
                phone: { type: string }
                address: { type: string }
                city: { type: string }
                state: { type: string }
                zip: { type: string }
                country: { type: string, default: "US" }
                website: { type: string }
                notes: { type: string }
      responses:
        "201": { content: { application/json: { schema: { $ref: "#/components/schemas/Company" } } } }

  # ─── Billing ───
  /billing/agreements:
    get:
      tags: [Billing]
      summary: List service agreements
      security: [{ bearerAuth: [] }]
      responses:
        "200": { content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/ServiceAgreement" } } } } }
    post:
      tags: [Billing]
      summary: Create a service agreement
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, companyId]
              properties:
                name: { type: string }
                companyId: { type: string }
                description: { type: string }
                billingPeriod: { $ref: "#/components/schemas/BillingPeriod" }
                price: { type: number }
                startDate: { type: string, format: date }
                endDate: { type: string, format: date }
                autoRenew: { type: boolean }
                cancellationDays: { type: integer }
      responses:
        "201": { content: { application/json: { schema: { $ref: "#/components/schemas/ServiceAgreement" } } } }
  /billing/invoices:
    get:
      tags: [Billing]
      summary: List invoices
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/InvoiceStatus" } }
        - { name: companyId, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Paginated invoice list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Invoice" } }
                  total: { type: integer }
  /billing/invoices/generate:
    post:
      tags: [Billing]
      summary: Auto-generate invoice from unbilled time entries
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [companyId]
              properties:
                companyId: { type: string }
                agreementId: { type: string }
      responses:
        "201": { content: { application/json: { schema: { $ref: "#/components/schemas/Invoice" } } } }
  /billing/invoices/{id}/send:
    post:
      tags: [Billing]
      summary: Send invoice to client via email
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: "Invoice sent" }
  /billing/invoices/{id}/record-payment:
    post:
      tags: [Billing]
      summary: Record a payment against an invoice
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: number }
                method: { type: string }
                reference: { type: string }
      responses:
        "200": { description: "Payment recorded" }

  # ─── Integrations ───
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InvoiceStatus:
      type: string
      enum: [draft, sent, partial, paid, overdue, void, collections]
    BillingPeriod:
      type: string
      enum: [one_time, weekly, monthly, quarterly, semi_annually, annually]
...
    Board:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        enabled: { type: boolean }
        autoCloseDays: { type: integer }
        createdAt: { type: string, format: date-time }
    BoardDetail:
      allOf:
        - $ref: "#/components/schemas/Board"
        - type: object
          properties:
            _count: { type: object, properties: { tickets: { type: integer } } }
            emailConnectors: { type: array, items: { $ref: "#/components/schemas/EmailConnector" } }
    EmailConnector:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        host: { type: string }
        port: { type: integer }
        secure: { type: boolean }
        folder: { type: string }
        pollIntervalSeconds: { type: integer }
        enabled: { type: boolean }
        lastCheckedAt: { type: string, format: date-time }
    Company:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        address: { type: string }
        city: { type: string }
        state: { type: string }
        zip: { type: string }
        country: { type: string }
        website: { type: string }
        notes: { type: string }
        status: { type: string }
    ServiceAgreement:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        companyId: { type: string }
        description: { type: string }
        billingPeriod: { $ref: "#/components/schemas/BillingPeriod" }
        price: { type: number }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        autoRenew: { type: boolean }
        cancellationDays: { type: integer }
        status: { type: string }
    Invoice:
      type: object
      properties:
        id: { type: string }
        invoiceNumber: { type: string }
        companyId: { type: string }
        agreementId: { type: string }
        issueDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        subtotal: { type: number }
        taxRate: { type: number }
        taxTotal: { type: number }
        total: { type: number }
        status: { $ref: "#/components/schemas/InvoiceStatus" }
        sentAt: { type: string, format: date-time }
        paidAt: { type: string, format: date-time }
        lineItems: { type: array, items: { $ref: "#/components/schemas/InvoiceLineItem" } }
    InvoiceLineItem:
      type: object
      properties:
        id: { type: string }
        description: { type: string }
        quantity: { type: number }
        unitPrice: { type: number }
        total: { type: number }
...